refactor: adjusting the Service methods

+++ src/main/java/com/example/todolist/service/TodoService.java
@@ -1,12 +1,14 @@
 package com.example.todolist.service;
 
 import com.example.todolist.domain.Todo;
+import com.example.todolist.exception.TodoNotFoundException;
 import com.example.todolist.repository.TodoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TodoService {
@@ -20,18 +22,25 @@ public class TodoService {
         return todoRepository.findAll(listaOrdenada);
     }
 
-    public List<Todo> createToDo(Todo todo) {
-        todoRepository.save(todo);
-        return listAll();
+    public Todo createTodo(Todo todo) {
+        return todoRepository.save(todo);
     }
 
-    public List<Todo> updateToDo(Todo todo) {
-        todoRepository.save(todo);
-        return listAll();
+    public Todo updateTodo(Todo todo) {
+        if (!todoRepository.existsById(todo.getId())) {
+            throw new TodoNotFoundException("Todo não encontrada com o id " + todo.getId());
+        }
+        return todoRepository.save(todo);
     }
 
-    public List<Todo> deleteTodo(Long id) {
+    public void deleteTodo(Long id) {
+        if(!todoRepository.existsById(id)) {
+            throw new TodoNotFoundException("Todo não encontrada com o id " + id);
+        }
         todoRepository.deleteById(id);
-        return listAll();
+    }
+
+    public Optional<Todo> findById(Long id) {
+        return todoRepository.findById(id);
     }
 }
